Resolves #34 

I've run test benchmark on 1024x1024 bitmap with a circle of 384px radius as fill target.

All the changes mentioned in Reddit post are implemented.

The only other significant performance boost would be to parallelize bitmap traversal, but since in WPF bitmaps have thread affinity, it would require cloning whole bitmap...

### Before

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1082 (1909/November2018Update/19H2)
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.402
  [Host]        : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  .NET Core 3.1 : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT

Job=.NET Core 3.1  Runtime=.NET Core 3.1  

```
|        Method |    Mean |    Error |   StdDev |     Min |     Max |  Median |
|-------------- |--------:|---------:|---------:|--------:|--------:|--------:|
| FloodFillTest | 1.298 s | 0.0038 s | 0.0036 s | 1.291 s | 1.304 s | 1.298 s |

### After

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1082 (1909/November2018Update/19H2)
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.402
  [Host]        : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  .NET Core 3.1 : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT

Job=.NET Core 3.1  Runtime=.NET Core 3.1  

```
|        Method |    Mean |    Error |   StdDev |     Min |     Max |  Median |
|-------------- |--------:|---------:|---------:|--------:|--------:|--------:|
| FloodFillTest | 1.016 s | 0.0064 s | 0.0057 s | 1.005 s | 1.027 s | 1.016 s |